Update fullpost.php
Fullpost.php allows for the user to view the shoes individually if content is over 200 characters or by searching for a certain shoe. Since shoe size is a new column addition to the database I added it to fullpost.php so users can see the size of the shoe after a search or if they are viewing the full post.