Update index.php
header.php is the search bar and button connecting to the search.php. I decided to make it a seperate php file so its easier to read and a lot more organized instead of adding a whole search section to the index page. This allows users to search for a shoe on the index page.