Update search.php
This is a new file called search.php it is a separate page after a user searches for a shoe, it will direct them to this PHP page displaying the main header of the website and the footer to access the other pages. Mainly this shows the actual results of the search made by the user to indicate which keywords matched that of the characters the user typed into the search bar. If results match any characters the links to the shoes will be displayed and the user can access these shoes one by one due to the linked fullpost.php file attached to each true result.